Added DB Session Handler
